FileSelector component
Moved the file select functionality into a separate and reusable component
