+++ frontend/src/components/Inputs/FileSelector.vue
@@ -1,0 +1,56 @@
+<template>
+  <div class="d-flex">
+    <input type="text"
+      class="form-control mr-1"
+      :disabled="true"
+      aria-describedby="basic-addon2"
+      :value=value>
+
+    <p-button v-if="action == 'SelectFile'"
+      @click.native="importKey" 
+      type="default">
+      <span style="display: block;">
+        BROWSE
+      </span>
+    </p-button>
+    <p-button v-if="action == 'SelectSaveFile'"
+      @click.native="SelectDirToStoreKey" 
+      type="default">
+      <span style="display: block;">
+        BROWSE
+      </span>
+    </p-button>
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    value: String,
+    placeholder: String,
+    action: String
+  },
+
+  methods: {
+    importKey: function() {
+      window.backend.WalletApplication.ImportKey().then(
+        result => {
+          if (result) {
+            this.$store.state.walletInfo.keystorePath = result;
+          }
+        }
+      );
+    },
+    SelectDirToStoreKey: function() {
+      window.backend.WalletApplication.SelectDirToStoreKey().then(
+        result => {
+          this.$store.state.walletInfo.saveKeystorePath = result;
+          this.$store.state.walletInfo.keystorePath = result;
+        }
+      );
+    }
+  }
+}
+</script>
+
+<style scoped></style>
