+++ frontend/src/plugins/globalComponents.js
@@ -1,4 +1,4 @@
-import { FormGroupInput, Card, DropDown, Button, Password } from "../components/index";
+import { FormGroupInput, Card, DropDown, Button, Password, FileSelector } from "../components/index";
 
 /**
  * You can register global components here and use them as a plugin in your main Vue instance
@@ -11,6 +11,7 @@ const GlobalComponents = {
     Vue.component("card", Card);
     Vue.component("p-button", Button);
     Vue.component("password-input", Password);
+    Vue.component("file-selector", FileSelector);
   }
 };
